float to int

+++ Example_0_2_Random_Distribution/src/ofApp.h
@@ -23,6 +23,6 @@ public:
   void dragEvent(ofDragInfo dragInfo) override;
   void gotMessage(ofMessage msg) override;
 
-  vector<float> randomCounts;
+  vector<int> randomCounts;
   int total;
 };
